+++ src/ncea-harvester.tests/Processors/JnccProcessorTests.cs
@@ -2,6 +2,7 @@
 using Azure.Storage.Blobs;
 using Microsoft.Extensions.Logging;
 using Moq;
+using ncea.harvester.Services;
 using Ncea.Harvester.BusinessExceptions;
 using Ncea.Harvester.Constants;
 using Ncea.Harvester.Models;
@@ -20,23 +21,6 @@ public class JnccProcessorTests
     {
         _mockOrchestrationServiceLogger = new Mock<ILogger<OrchestrationService>>(MockBehavior.Strict);
         _mockLogger = new Mock<ILogger<JnccProcessor>>(MockBehavior.Strict);
-
-        //_mockOrchestrationServiceLogger.Setup(x => x.Log(
-        //        LogLevel.Information,
-        //        It.IsAny<EventId>(),
-        //        It.IsAny<It.IsAnyType>(),
-        //        It.IsAny<Exception>(),
-        //        It.IsAny<Func<It.IsAnyType, Exception?, string>>()
-        //    )
-        //);
-        //_mockOrchestrationServiceLogger.Setup(x => x.Log(
-        //        LogLevel.Error,
-        //        It.IsAny<EventId>(),
-        //        It.IsAny<It.IsAnyType>(),
-        //        It.IsAny<Exception>(),
-        //        It.IsAny<Func<It.IsAnyType, Exception?, string>>()
-        //    )
-        //);
         _mockLogger.Setup(x => x.Log(
                 LogLevel.Information,
                 It.IsAny<EventId>(),
@@ -205,86 +189,7 @@ public class JnccProcessorTests
             Times.Exactly(2),
             It.IsAny<string>()
         );
-    }
-
-    //[Fact]
-    //public async Task Process_WhenServiceBusThrowsException_ShouldLogMessage()
-    //{
-    //    //Arrange
-    //    var serviceBusService = ServiceBusServiceForTests.GetServiceBusWithError(out Mock<ServiceBusSender> mockServiceBusSender);
-    //    var metaDataXmlStr = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n\r\n<gmd:MD_Metadata\r\n        xmlns:gmd=\"http://www.isotc211.org/2005/gmd\"\r\n        xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"\r\n        xmlns:gml=\"http://www.opengis.net/gml\" xmlns:gts=\"http://www.isotc211.org/2005/gts\"\r\n        xmlns:mdc=\"https://github.com/DEFRA/ncea-geonetwork/tree/main/core-geonetwork/schemas/iso19139.mdc/src/main/plugin/iso19139.mdc/schema/mdc\"\r\n        xmlns:gco=\"http://www.isotc211.org/2005/gco\">\r\n  <gmd:fileIdentifier>\r\n    <gco:CharacterString>8b1fd363-cfed-49f0-b6e2-8eab3138a735</gco:CharacterString>\r\n  </gmd:fileIdentifier></gmd:MD_Metadata>";
-    //    var expectedData = "<html><body><a href=\"a.xml\">a</a><a href=\"b.xml\">b</a></body></html>";
-    //    var httpResponse = new HttpResponseMessage
-    //    {
-    //        StatusCode = HttpStatusCode.OK,
-    //        Content = new StringContent(expectedData),
-    //    };
-    //    var apiClient = ApiClientForTests.Get(httpResponse);
-    //    var harvesterConfiguration = new HarvesterConfiguration() { DataSourceApiBase = "https://base-uri", DataSourceApiUrl = "/test-url", ProcessorType = ProcessorType.Jncc, Type = "" };
-
-    //    var blobService = BlobServiceForTests.Get(out Mock<BlobServiceClient> mockBlobServiceClient,
-    //                                          out Mock<BlobContainerClient> mockBlobContainerClient,
-    //                                          out Mock<BlobClient> mockBlobClient);
-
-    //    var orchestrationservice = new OrchestrationService(blobService, serviceBusService, _mockOrchestrationServiceLogger.Object);
-
-    //    // Act
-    //    var jnccMockService = new Mock<JnccProcessor>(apiClient, orchestrationservice, _mockLogger.Object, harvesterConfiguration);
-    //    jnccMockService.Setup(x => x.GetJnccMetadata(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Returns(Task.FromResult(metaDataXmlStr));
-    //    await jnccMockService.Object.ProcessAsync(It.IsAny<CancellationToken>());
-
-    //    // Assert
-    //    _mockOrchestrationServiceLogger.Verify(
-    //        m => m.Log(
-    //            LogLevel.Error,
-    //            It.IsAny<EventId>(),
-    //            It.IsAny<It.IsAnyType>(),
-    //            It.IsAny<Exception>(),
-    //            It.IsAny<Func<It.IsAnyType, Exception?, string>>()),
-    //        Times.Exactly(2),
-    //        It.IsAny<string>()
-    //    );
-    //}
-
-    //[Fact]
-    //public async Task Process_WhenBlobThrowsException_ShouldLogMessage()
-    //{
-    //    //Arrange
-    //    var serviceBusService = ServiceBusServiceForTests.Get(out Mock<ServiceBusSender> mockServiceBusSender);
-    //    var metaDataXmlStr = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n\r\n<gmd:MD_Metadata\r\n        xmlns:gmd=\"http://www.isotc211.org/2005/gmd\"\r\n        xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"\r\n        xmlns:gml=\"http://www.opengis.net/gml\" xmlns:gts=\"http://www.isotc211.org/2005/gts\"\r\n        xmlns:mdc=\"https://github.com/DEFRA/ncea-geonetwork/tree/main/core-geonetwork/schemas/iso19139.mdc/src/main/plugin/iso19139.mdc/schema/mdc\"\r\n        xmlns:gco=\"http://www.isotc211.org/2005/gco\">\r\n  <gmd:fileIdentifier>\r\n    <gco:CharacterString>8b1fd363-cfed-49f0-b6e2-8eab3138a735</gco:CharacterString>\r\n  </gmd:fileIdentifier></gmd:MD_Metadata>";
-    //    var expectedData = "<html><body><a href=\"a.xml\">a</a><a href=\"b.xml\">b</a></body></html>";
-    //    var httpResponse = new HttpResponseMessage
-    //    {
-    //        StatusCode = HttpStatusCode.OK,
-    //        Content = new StringContent(expectedData),
-    //    };
-    //    var apiClient = ApiClientForTests.Get(httpResponse);
-    //    var harvesterConfiguration = new HarvesterConfiguration() { DataSourceApiBase = "https://base-uri", DataSourceApiUrl = "/test-url", ProcessorType = ProcessorType.Jncc, Type = "" };
-
-    //    var blobService = BlobServiceForTests.GetWithError(out Mock<BlobServiceClient> mockBlobServiceClient,
-    //                                          out Mock<BlobContainerClient> mockBlobContainerClient,
-    //                                          out Mock<BlobClient> mockBlobClient);        
-        
-    //    var orchestrationservice = new OrchestrationService(blobService, serviceBusService, _mockOrchestrationServiceLogger.Object);
-
-    //    // Act
-    //    var jnccMockService = new Mock<JnccProcessor>(apiClient, orchestrationservice, _mockLogger.Object, harvesterConfiguration);
-    //    jnccMockService.Setup(x => x.GetJnccMetadata(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Returns(Task.FromResult(metaDataXmlStr));
-    //    await jnccMockService.Object.ProcessAsync(It.IsAny<CancellationToken>());
-
-    //    // Assert
-    //    _mockOrchestrationServiceLogger.Verify(
-    //        m => m.Log(
-    //            LogLevel.Error,
-    //            It.IsAny<EventId>(),
-    //            It.IsAny<It.IsAnyType>(),
-    //            It.IsAny<Exception>(),
-    //            It.IsAny<Func<It.IsAnyType, Exception?, string>>()),
-    //        Times.Exactly(2),
-    //        It.IsAny<string>()
-    //    );
-    //}
-
+    }    
 
     [Fact]
     public async Task Process_WhenJnccMetaDataApiCallThrowsHttpException_ShouldThrowError()
