+++ src/ncea-harvester/Program.cs
@@ -13,7 +13,8 @@ using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Logging.ApplicationInsights;
 using Microsoft.ApplicationInsights.DependencyCollector;
 using Microsoft.Extensions.Azure;
-using Ncea.Harvester.Processors;
+using ncea.harvester.Services.Contracts;
+using ncea.harvester.Services;
 
 var configuration = new ConfigurationBuilder()
                                 .SetBasePath(Directory.GetCurrentDirectory())
